+++ Src/SOS.Observations.Api/Controllers/ExportsController.cs
@@ -11,6 +11,7 @@ using Microsoft.Extensions.Logging;
 using SOS.Lib.Configuration.ObservationApi;
 using SOS.Lib.Constants;
 using SOS.Lib.Enums;
+using SOS.Lib.Extensions;
 using SOS.Lib.Helpers;
 using SOS.Lib.Jobs.Export;
 using SOS.Lib.Managers.Interfaces;
@@ -37,109 +38,6 @@ namespace SOS.Observations.Api.Controllers
         private readonly long _downloadExportObservationsLimit;
         private readonly string _exportPath;
         private readonly ILogger<ExportsController> _logger;
-
-        /// <summary>
-        /// Populate output fields based on property set
-        /// </summary>
-        /// <param name="filter"></param>
-        /// <param name="exportPropertySet"></param>
-        private void PopulateOutputFields(SearchFilter filter, ExportPropertySet exportPropertySet)
-        {
-            if (exportPropertySet == ExportPropertySet.All)
-            {
-                return;
-            }
-
-            var outputFields = new List<string>
-            {
-                "DatasetName",
-                "Identification.Validated",
-                "Identification.UncertainIdentification",
-                "Location.County.Name",
-                "Location.Municipality.Name",
-                "Location.DecimalLongitude",
-                "Location.DecimalLatitude",
-                "Location.CoordinateUncertaintyInMeters",
-                "Occurrence.OccurrenceStatus.Value",
-                "Occurrence.RecordedBy",
-                "Taxon.Attributes.OrganismGroup",
-                "Taxon.ScientificName",
-                "Taxon.VernacularName"
-            };
-
-            if (exportPropertySet == ExportPropertySet.Extended)
-            {
-                outputFields.AddRange(new []
-                {
-                    "CollectionCode",
-                    "InstitutionCode",
-                    "OwnerInstitutionCode",
-                    "BasisOfRecord",
-                    "Event.Habitat",
-                    "Event.EventRemarks",
-                    "Event.SamplingEffort",
-                    "Event.SamplingProtocol",
-                    "Event.SampleSizeUnit",
-                    "Event.SampleSizeValue",
-                    "Location.Locality",
-                    "Location.Province.Name",
-                    "Location.Parish.Name",
-                    "Location.GeodeticDatum",
-                    "Occurrence.ReportedBy",
-                    "Occurrence.Url",
-                    "Occurrence.AssociatedMedia",
-                    "Occurrence.OccurrenceRemarks",
-                    "Occurrence.Activity.Value",
-                    "Occurrence.Behavior.Value",
-                    "Occurrence.LifeStage.Value",
-                    "Occurrence.ReproductiveCondition.Value",
-                    "Occurrence.Sex.Value",
-                    "Occurrence.Biotope.Value",
-                    "Occurrence.BiotopeDescription",
-                    "Occurrence.ProtectionLevel",
-                    "Occurrence.IsNeverFoundObservation",
-                    "Occurrence.IsNotRediscoveredObservation",
-                    "Occurrence.IsNaturalOccurrence",
-                    "Occurrence.IsPositiveObservation",
-                    "Occurrence.Substrate.Name.Value",
-                    "Occurrence.individualCount",
-                    "Occurrence.OrganismQuantity",
-                    "Occurrence.OrganismQuantityInt",
-                    "Occurrence.OrganismQuantityUnit",
-                    "Identification.ValidationStatus",
-                    "Identification.ConfirmedBy",
-                    "Identification.IdentifiedBy",
-                    "Identification.VerifiedBy",
-                    "Identification.DeterminationMethod.Value",
-                    "Taxon.Kingdom",
-                    "Taxon.Phylum",
-                    "Taxon.Class",
-                    "Taxon.Order",
-                    "Taxon.Family",
-                    "Taxon.Genus",
-                    "Taxon.TaxonId",
-                    "Taxon.Attributes.DyntaxaTaxonId",
-                    "Taxon.Attributes.ProtectionLevel.Value",
-                    "Taxon.Attributes.RedlistCategory",
-                    "Taxon.Attributes.ProtectedByLaw",
-                    "Project.Id",
-                    "Project.Name"
-                });
-            }
-
-            // Order fields by name
-            outputFields = outputFields.OrderBy(s => s).ToList();
-
-            // Make sure some selected fields occurs first
-            outputFields.InsertRange(0, new []
-            {
-                "Occurrence.OccurrenceId",
-                "Event.StartDate",
-                "Event.EndDate"
-            });
-
-            filter.OutputFields = outputFields;
-        }
 
         /// <summary>
         /// Validate input for order request
@@ -308,7 +206,7 @@ namespace SOS.Observations.Api.Controllers
         [HttpPost("Download/Excel")]
         [ProducesResponseType(typeof(byte[]), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
-        public async Task<IActionResult> DownloadExcelAsync([FromBody] ExportFilterDto filter, [FromQuery] ExportPropertySet exportPropertySet, [FromQuery] string cultureCode)
+        public async Task<IActionResult> DownloadExcelAsync([FromBody] ExportFilterDto filter, [FromQuery] OutputFieldSet outputFieldSet, [FromQuery] string cultureCode)
         {
             cultureCode = CultureCodeHelper.GetCultureCode(cultureCode);
             var filePath = string.Empty;
@@ -322,8 +220,8 @@ namespace SOS.Observations.Api.Controllers
                 }
 
                 var exportFilter = (SearchFilter)okResult.Value;
-                PopulateOutputFields(exportFilter, exportPropertySet);
-
+                exportFilter.PopulateOutputFields(outputFieldSet);
+               
                 filePath =
                     await _exportManager.CreateExportFileAsync(exportFilter, ExportFormat.Excel, 
                         _exportPath, cultureCode,
@@ -347,7 +245,7 @@ namespace SOS.Observations.Api.Controllers
         [ProducesResponseType(typeof(byte[]), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> DownloadGeoJsonAsync([FromBody] ExportFilterDto filter, 
-            [FromQuery] ExportPropertySet exportPropertySet, [FromQuery] string cultureCode, [FromQuery] bool flatOut)
+            [FromQuery] OutputFieldSet outputFieldSet, [FromQuery] string cultureCode, [FromQuery] bool flatOut)
         {
             cultureCode = CultureCodeHelper.GetCultureCode(cultureCode);
             var filePath = string.Empty;
@@ -361,7 +259,7 @@ namespace SOS.Observations.Api.Controllers
                 }
 
                 var exportFilter = (SearchFilter)okResult.Value;
-                PopulateOutputFields(exportFilter, exportPropertySet);
+                exportFilter.PopulateOutputFields(outputFieldSet);
 
                 filePath =
                     await _exportManager.CreateExportFileAsync(exportFilter, ExportFormat.GeoJson, 
@@ -418,7 +316,7 @@ namespace SOS.Observations.Api.Controllers
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> OrderExcelAsync([FromBody] ExportFilterDto filter, [FromQuery] string description, 
-            [FromQuery] ExportPropertySet exportPropertySet, [FromQuery] string cultureCode)
+            [FromQuery] OutputFieldSet outputFieldSet, [FromQuery] string cultureCode)
         {
             try
             {
@@ -431,7 +329,7 @@ namespace SOS.Observations.Api.Controllers
                 }
 
                 var (email, exportFilter) = ((string, SearchFilter))okResult.Value;
-                PopulateOutputFields(exportFilter, exportPropertySet);
+                exportFilter.PopulateOutputFields(outputFieldSet);
 
                 return new OkObjectResult(BackgroundJob.Enqueue<IExportAndSendJob>(job =>
                     job.RunAsync(exportFilter, email, description, ExportFormat.Excel, cultureCode, false, JobCancellationToken.Null)));
@@ -451,7 +349,7 @@ namespace SOS.Observations.Api.Controllers
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> OrderGeoJsonAsync([FromBody] ExportFilterDto filter, [FromQuery] string description, 
-            [FromQuery] ExportPropertySet exportPropertySet, [FromQuery] string cultureCode, [FromQuery] bool flatOut)
+            [FromQuery] OutputFieldSet outputFieldSet, [FromQuery] string cultureCode, [FromQuery] bool flatOut)
         {
             try
             {
@@ -464,7 +362,7 @@ namespace SOS.Observations.Api.Controllers
                 }
 
                 var (email, exportFilter) = ((string, SearchFilter))okResult.Value;
-                PopulateOutputFields(exportFilter, exportPropertySet);
+                exportFilter.PopulateOutputFields(outputFieldSet);
 
                 return new OkObjectResult(BackgroundJob.Enqueue<IExportAndSendJob>(job =>
                     job.RunAsync(exportFilter, email, description, ExportFormat.GeoJson, cultureCode, flatOut, JobCancellationToken.Null)));